backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM & AI (Updated for compatibility)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0
tavily-python>=0.3.0

# Vector Store & Embeddings
chromadb>=0.4.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# HTTP Clients
httpx>=0.26.0
requests>=2.31.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Rate Limiting
slowapi>=0.1.9

# Prometheus Metrics
prometheus-client>=0.19.0

# Utilities
python-dateutil>=2.8.0
